+++ datasource/netcdf_groundwater.py
@@ -63,32 +63,52 @@ def find_aggregation_netcdf_gw(netcdf_file_path):
     return glob.glob(os.path.join(result_dir, pattern))[0]
 
 
-class NetcdfDataSourceGroundwater(BaseDataSource):
+class NetcdfGroundwaterDataSource(BaseDataSource):
     PREFIX_1D = 'Mesh1D_'
     PREFIX_2D = 'Mesh2D_'
     PREFIX_1D_LENGTH = 7  # just so we don't have to recalculate
     PREFIX_2D_LENGTH = 7  # just so we don't have to recalculate
 
     def __init__(self, file_path=None, *args, **kwargs):
-        from netCDF4 import Dataset
         self.file_path = file_path
         self._ga = None
         self._ga_result = None
-        self.ds = Dataset(file_path)
-
-        self.nMesh2D_nodes = self.ds.dimensions['nMesh2D_nodes'].size
-        self.nMesh1D_nodes = self.ds.dimensions['nMesh1D_nodes'].size
-        self.nMesh2D_lines = self.ds.dimensions['nMesh2D_lines'].size
-        self.nMesh1D_lines = self.ds.dimensions['nMesh1D_lines'].size
+        self._ds = None
         self._cache = {}
+
+    @property
+    def ds(self):
+        from netCDF4 import Dataset
+        if self._ds is None:
+            try:
+                self._ds = Dataset(self.file_path)
+            except IOError:
+                pass
+        return self._ds
+
+    @property
+    def nMesh2D_nodes(self):
+        return self.ds.dimensions['nMesh2D_nodes'].size
+
+    @property
+    def nMesh1D_nodes(self):
+        return self.ds.dimensions['nMesh1D_nodes'].size
+
+    @property
+    def nMesh2D_lines(self):
+        return self.ds.dimensions['nMesh2D_lines'].size
+
+    @property
+    def nMesh1D_lines(self):
+        return self.ds.dimensions['nMesh1D_lines'].size
 
     def _strip_prefix(self, var_name):
         """Strip away netCDF variable name prefixes.
 
         Example variable names: 'Mesh2D_s1', 'Mesh1D_s1'
 
-        >>> from ThreeDiToolbox.datasource.netcdf_groundwater import NetcdfDataSourceGroundwater  # noqa
-        >>> ds = NetcdfDataSourceGroundwater()
+        >>> from ThreeDiToolbox.datasource.netcdf_groundwater import NetcdfGroundwaterDataSource  # noqa
+        >>> ds = NetcdfGroundwaterDataSource()
         >>> ds._strip_prefix('Mesh2D_s1')
         's1'
         >> ds._strip_prefix('Mesh1D_q')